Substantial update to browser icon behaviour (extra badge feedback) and JS-set cookies
Icons updated courtesy of ArcticGems, and now the popup icon shows whether cookies were allowed/blocked, for first and third party.
Realised as part of that work that iframes really needed to be called out, and now they are (in the popup).
Also found a much better way to detect cookies set *from* javascript, and was able to remove huge chunks of super dodgy code that guessed the actual behaviour.