Make processing javascript cookies a bit more and less reliable
Cookies set in <script> in the body of documents will be set probably (but not always) before the config has loaded in the content script, meaning the 'set' implementation failed hard.  This commit rewrites that to do the initial config load and 'eval' injection inside an async function that is called at load time; it means cookies might sneak by the content script, but at least the capturing won't fail hard and utterly prevent the cookie from being set at all ever.

This does not address attempting to capture those early cookies, which is still to be done.  There are also other changes that could be done (see the TODO committed) to make this somewhat better (although still potentially racy), but which will require more effort.  The event-based capturing of cookies is still required for other reasons (CSP), and will be tackled first