Use a script instead of an eval to inject cookie hooks
Content-Security-Policy (CSP) headers were the downfall for the eval; they prevented the hook ever working in that case, so we were always doing fallback handling (catch-all on the cookie-set event) in background.js, which was functional but deeply sub-optimal (we stopped tracking which tab that came from a while ago, because it was so janky).

We could add 'unsafe-eval' to the CSP header but that would be borderline evil.  With a script tag added to the document, we can (and do) add a sha256 hash directive to any script-src portions of a CSP header, meaning we've only allowed exactly our script to run in addition to whatever the page author permits.  It's still not ideal to be munging CSP (it's lightly risky, although all due care is taken), but this does work.

It's still not quite as good as a straight up eval in terms of getting in early enough at page load (because the tag will inherently be evaluated 'later', and perhaps after some cookies are set by javascript that loads earlier), but we have backup fallback handling in background.js (see above) *and* the eval had to be async anyway due to needing to async load 'config'.

So this is better than eval in many ways (CSP), and not substantially worse (timing).