Fixes https://github.com/craigmiskell/cookiemaster/issues/18
Oh the embarrassment.  Fixes a failure to delete cookies when javascript tried to (by setting an expires value to "before now", or a negative max-age).  Did assume that just 'not doing anything' was sufficient, but we still needed to pass the 'delete' cookie definition to document.cookie for the deletion to be processed.

Probably because in the code it was derived from (in background.js) that *was* the correct thing to do (return, do nothing, which let the header filtering code carry on as though nothing was changed).

Interfaces are important; bugs come from failures to understand interfaces between components.